refactor: 기능별 / 담당하는 내용별로 패키지를 분리
