+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".MovieActivity"
+            android:name=".view.MovieActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -21,8 +21,8 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".MovieBookingActivity" />
-        <activity android:name=".MovieBookedActivity" />
+        <activity android:name=".view.MovieBookingActivity" />
+        <activity android:name=".view.MovieBookedActivity" />
     </application>
 
 </manifest>
