+++ app/src/main/res/layout/movie_booked.xml
@@ -5,7 +5,7 @@
     android:id="@+id/booked"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MovieBookedActivity">
+    tools:context=".view.MovieBookedActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/notice_text_container"
