+++ app/src/main/res/layout/movie_booking.xml
@@ -5,7 +5,7 @@
     android:id="@+id/booking"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MovieActivity">
+    tools:context=".view.MovieActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/notice_text_container"
